refactor: useCurrentUserLazyQuery
