+++ apps/journeys-admin/src/components/Team/TeamManageDialog/TeamManageDialog.spec.tsx
@@ -18,9 +18,9 @@ jest.mock('@mui/material/useMediaQuery', () => ({
   default: jest.fn()
 }))
 
-jest.mock('apps/journeys-admin/src/libs/useCurrentUser', () => ({
+jest.mock('apps/journeys-admin/src/libs/useCurrentUserLazyQuery', () => ({
   __esModule: true,
-  useCurrentUser: jest.fn().mockReturnValue({
+  useCurrentUserLazyQuery: jest.fn().mockReturnValue({
     loadUser: jest.fn(),
     data: {
       __typename: 'User',
