+++ apps/journeys-admin/src/components/Team/TeamManageDialog/TeamManageDialog.stories.tsx
@@ -4,7 +4,7 @@ import { Meta, StoryObj } from '@storybook/react'
 import { UserTeamRole } from '../../../../__generated__/globalTypes'
 import { ApolloLoadingProvider } from '../../../../test/ApolloLoadingProvider'
 import { journeysAdminConfig } from '../../../libs/storybook'
-import { GET_CURRENT_USER } from '../../../libs/useCurrentUser'
+import { GET_CURRENT_USER } from '../../../libs/useCurrentUserLazyQuery'
 import { GET_USER_TEAMS_AND_INVITES } from '../../../libs/useUserTeamsAndInvitesQuery/useUserTeamsAndInvitesQuery'
 import {
   GET_LAST_ACTIVE_TEAM_ID_AND_TEAMS,
