--- apps/journeys-admin/src/libs/useCurrentUser/index.ts
@@ -1,1 +1,0 @@
-export { useCurrentUser, GET_CURRENT_USER } from './useCurrentUser'
