+++ apps/journeys-admin/src/libs/useCurrentUserLazyQuery/index.ts
@@ -1,0 +1,4 @@
+export {
+  useCurrentUserLazyQuery,
+  GET_CURRENT_USER
+} from './useCurrentUserLazyQuery'
